reverted breaking change to rgb_to_hsl method

+++ panel.h
@@ -20,8 +20,7 @@ class Panel {
       setPixelAverage(uint16_t x, uint8_t y, uint8_t r, uint8_t g, uint8_t b),
       setPixelAverage(uint16_t x, uint8_t y, uint32_t c),
       color_to_hsl(uint32_t color, float * h, float * s, float * l),
-      //rgb_to_hsl(uint8_t r, uint8_t g, uint8_t b, float * h, float * s, float * l),
-      rgb_to_hsl(float r, float g, float b, float * h, float * s, float * l),
+      rgb_to_hsl(uint8_t r, uint8_t g, uint8_t b, float * h, float * s, float * l),
       hsl_to_rgb(float h, float s, float l, uint8_t * red, uint8_t * green, uint8_t * blue);
     double
       hue2rgb(double p, double q, float t);
